Refactor everything

+++ BlackBeans/Persistency/Persistency.swift
@@ -11,12 +11,43 @@ import CoreData
 import Combine
 
 protocol Persistable {
-  static func createBean(name: String, value: Decimal) -> AnyPublisher<Void, Error>
+  /// Returns the sum of all beans values
+  static var allBeansSum: Decimal { get }
+  
+  /// Returns a Fetch Request with all beans ordered by name
+  static var allBeansFetchRequest: NSFetchRequest<Bean> { get }
+  
+  /// Create a new bean
+  static func createBean(name: String, value: Decimal) throws
+  
+  /// Deletes a bean
+  static func deleteBean(bean: Bean?) throws
+  
+  /// Update existing bean
+  static func updateBean(bean: Bean, name: String, value: Decimal) throws
 }
 
 class Persistency: NSObject {
   
   private static let shared = Persistency()
+  
+  static var viewContext: NSManagedObjectContext {
+    return shared.persistentContainer.viewContext
+  }
+  
+  private lazy var persistentContainer: NSPersistentContainer = {
+      let container = NSPersistentContainer(name: "BlackBeans")
+      container.loadPersistentStores(completionHandler: { (storeDescription, error) in
+          if let error = error as NSError? {
+              fatalError("Unresolved error \(error), \(error.userInfo)")
+          }
+      })
+      return container
+  }()
+  
+}
+
+extension Persistency: Persistable {
   
   static var allBeansSum: Decimal {
     let expression = NSExpressionDescription()
@@ -34,50 +65,29 @@ class Persistency: NSObject {
     return (res?["sum"] ?? 0).decimalValue
   }
   
-  static var viewContext: NSManagedObjectContext {
-    return shared.persistentContainer.viewContext
+  static var allBeansFetchRequest: NSFetchRequest<Bean> {
+    let fetch = NSFetchRequest<Bean>(entityName: "Bean")
+    fetch.sortDescriptors = [NSSortDescriptor(key: "value", ascending: true)]
+    return fetch
   }
   
-  private lazy var persistentContainer: NSPersistentContainer = {
-      let container = NSPersistentContainer(name: "BlackBeans")
-      container.loadPersistentStores(completionHandler: { (storeDescription, error) in
-          if let error = error as NSError? {
-              fatalError("Unresolved error \(error), \(error.userInfo)")
-          }
-      })
-      return container
-  }()
-  
-  static private func saveContext() throws {
-    let context = Persistency.viewContext
-    if context.hasChanges {
-      try context.save()
-    }
+  static func createBean(name: String, value: Decimal) throws {
+    let bean = Bean(context: Persistency.viewContext)
+    bean.creationTimestamp = Date()
+    bean.name = name
+    bean.value = NSDecimalNumber(decimal: value)
+    try Persistency.viewContext.save()
   }
   
-  static func deleteBean(bean: Bean) throws {
+  static func deleteBean(bean: Bean?) throws {
+    guard let bean = bean else { return }
     Persistency.viewContext.delete(bean)
-    try saveContext()
+    try Persistency.viewContext.save()
   }
   
-}
-
-extension Persistency: Persistable {
-  
-  static func createBean(name: String, value: Decimal) -> AnyPublisher<Void, Error> {
-    Future<Void, Error> { promise in
-      let bean = Bean(context: Persistency.viewContext)
-      bean.name = name
-      bean.value = NSDecimalNumber(decimal: value)
-      bean.creationTimestamp = Date()
-      do {
-        try Persistency.saveContext()
-        promise(.success(()))
-      } catch {
-        Persistency.viewContext.delete(bean)
-        promise(.failure(error))
-      }
-    }.eraseToAnyPublisher()
+  static func updateBean(bean: Bean, name: String, value: Decimal) throws {
+    bean.name = name
+    bean.value = NSDecimalNumber(decimal: value)
+    try Persistency.viewContext.save()
   }
-  
 }
